Restructure pytorch example code
