+++ .drone.yml
@@ -73,7 +73,7 @@ steps:
       AWS_SECRET_ACCESS_KEY:
         from_secret: minio-secret
     commands:
-      - python3 lab/processes/pytorch_example/prepare_data.py
+      - python3 lab/processes/pytorch_example/prepare_data/main.py
     volumes:
         - name: shared-storage
           path: /shared-storage
@@ -86,7 +86,7 @@ steps:
       AWS_SECRET_ACCESS_KEY:
         from_secret: minio-secret
     commands:
-      - python3 lab/processes/pytorch_example/train_model.py
+      - python3 lab/processes/pytorch_example/train_model/main.py
     volumes:
         - name: shared-storage
           path: /shared-storage
@@ -99,7 +99,7 @@ steps:
       AWS_SECRET_ACCESS_KEY:
         from_secret: minio-secret
     commands:
-      - python3 lab/processes/pytorch_example/test_model.py
+      - python3 lab/processes/pytorch_example/test_model/main.py
     volumes:
         - name: shared-storage
           path: /shared-storage
